[mem_tool] size_t to long (!71)

+++ src/util/pdm_mem_tool.c
@@ -1250,7 +1250,7 @@
 
 void *PDM_MALLOC
 (
- size_t nb,
+ long nb,
  size_t size_type,
  const char *func_name,
  const char *file_name,
@@ -1290,7 +1290,7 @@ void *PDM_REALLOC
 (
  void *old_ptr,
 // void *new_ptr,
- size_t nb,
+ long nb,
  size_t size_type,
  const char *func_name,
  const char *file_name,
@@ -1342,7 +1342,7 @@ void *PDM_REALLOC
 
 void *PDM_CALLOC
 (
- size_t nb,
+ long nb,
  size_t size_type,
  const char *func_name,
  const char *file_name,
